Inferences From Fitted G-SuSiE/SuSiE Model

These functions access basic properties or draw inference from a fitted susie/gsusie model. The codes and descriptions below are copied from https://github.com/stephenslab/susieR/blob/master/R/susie_utils.R

Usage

gsusie_get_objective(res, last_only = TRUE, warning_tol = 1e-06)

gsusie_get_log_pseudo_variance(res, X = NULL, scale = TRUE)

gsusie_get_posterior_mean(res, prior_tol = 1e-09)

gsusie_get_posterior_sd(res, prior_tol = 1e-09)

gsusie_get_niter(res)

gsusie_get_prior_variance(res)

gsusie_get_lfsr(res)

gsusie_get_posterior_samples(res, num_samples)

gsusie_get_cs(
  res,
  X = NULL,
  Xcorr = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  dedup = TRUE,
  squared = FALSE,
  check_symmetric = TRUE,
  n_purity = 100
)

gsusie_get_pip(res, prune_by_cs = FALSE, prior_tol = 1e-09)

Arguments

res

A susie fit, an output from susie or gsusie.

last_only

If last_only = FALSE, return the ELBO from all iterations; otherwise return the ELBO from the last iteration only.

warning_tol

Warn if ELBO is decreasing by this tolerance level.

X

n by p matrix of values of p variables (covariates) in n samples. When provided, correlation between variables will be computed and used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.

scale

Boolean, indicates whether input X needs to be scaled or not.

prior_tol

Filter out effects having estimated prior variance smaller than this threshold.

num_samples

The number of draws from the posterior distribution.

Xcorr

p by p matrix of correlations between variables (covariates). When provided, it will be used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.

coverage

A number between 0 and 1 specifying desired coverage of each CS.

min_abs_corr

A "purity" threshold for the CS. Any CS that contains a pair of variables with correlation less than this threshold will be filtered out and not reported.

dedup

If dedup = TRUE, remove duplicate CSs.

squared

If squared = TRUE, report min, mean and median of squared correlation instead of the absolute correlation.

check_symmetric

If check_symmetric = TRUE, perform a check for symmetry of matrix Xcorr when Xcorr is provided (not NULL).

n_purity

The maximum number of credible set (CS) variables used in calculating the correlation (“purity”) statistic. When the number of variables included in the CS is greater than this number, the CS variables are randomly subsampled.

prune_by_cs

Whether or not to ignore single effects not in a reported CS when calculating PIP.

Value

gsusie_get_objective returns the evidence lower bound (ELBO) achieved by the fitted susie model and, optionally, at each iteration of the IBSS fitting procedure.cs

gsusie_get_prior_variance returns the (estimated or fixed) prior variance parameters.

gsusie_get_posterior_mean returns the posterior mean for the regression coefficients of the fitted susie model.

gsusie_get_posterior_sd returns the posterior standard deviation for coefficients of the fitted susie model.

gsusie_get_niter returns the number of model fitting iterations performed.

gsusie_get_pip returns a vector containing the posterior inclusion probabilities (PIPs) for all variables.

gsusie_get_lfsr returns a vector containing the average lfsr across variables for each single-effect, weighted by the posterior inclusion probability (alpha).

gsusie_get_posterior_samples returns a list containing the effect sizes samples and causal status with two components: b, an num_variables x num_samples matrix of effect sizes; gamma, an num_variables x num_samples

matrix of causal status random draws.

gsusie_get_cs returns credible sets (CSs) from a susie fit, as well as summaries of correlation among the variables included in each CS. If desired, one can filter out CSs that do not meet a specified “purity” threshold; to do this, either X or Xcorr must be supplied. It returns a list with the following elements:

cs

A list in which each list element is a vector containing the indices of the variables in the CS.

coverage

The nominal coverage specified for each CS.

purity

If X or Xcorr iis provided), the purity of each CS.

cs_index

If X or Xcorr is provided) the index (number between 1 and L) of each reported CS in the supplied susie fit.