G-SuSiE Plots.

gsusie_plot produces a per-variable summary of the G-SuSiE credible sets.

Usage

gsusie_plot(
  model,
  y,
  include_intercept = FALSE,
  intercept_index = NULL,
  effect_indices = NULL,
  add_bar = FALSE,
  pos = NULL,
  max_cs = 400,
  add_legend = NULL,
  ...
)

Arguments

model

A SuSiE or G-SuSiE fit, typically an output from gsusie or one of its variants. For gsuse_plot, the susie fit must have model$z(?), model$PIP, and may include model$sets. model may also be a vector of z-scores or PIPs.

y

A string indicating what to plot: either "z_original" for z-scores, "z" for z-score derived p-values on (base-10) log-scale, "PIP" for posterior inclusion probabilities, "log10PIP" for posterior inclusion probabiliities on the (base-10) log-scale. For any other setting, the data are plotted as is.

include_intercept

Boolean. If include_intercept = FALSE, the intercept term is removed from the plot; the index of the intercept is specified in intercept_index.

intercept_index

a numeric number specifying the index of the intercept. It is required when include_intercept=FALSE (by default).

effect_indices

a numeric vector of indices of effect variables. Those points are highlighted in red.

add_bar

If add_bar = TRUE, add horizontal bar to signals in credible interval.

pos

This can be either be (1) a numeric vector of indices of subset of variables to plot, or (2) a list with the following list elements: pos$attr, pos$start and pos$end, where pos$attr is a character string of the name of index variable in model object, and pos$start and pos$end are boundaries of indices to plot. See the provided examples.

max_cs

The largest credible set to display, either based on purity (set max_cs between 0 and 1), or based on size (set max_cs > 1).

add_legend

If add_legend = TRUE, add a legend to annotate the size and purity of each CS discovered. It can also be specified as location where legends should be added, e.g., add_legend = "bottomright" (default location is "topright").

...

Additional arguments passed to plot.

Value

Invisibly returns NULL.