Robust estimations

When fitting a generalized linear model, outliers, i.e. unexpected values, may occur in both observations and intermediate procedures, such as weights (i.e. inverse of pseudo-variance) and pseudo-responses. Hence, robust estimations are be considered to "down-weigh" the influential points.

Usage

robust_importance_weights(
  values,
  robust_method = c("none", "huber", "simple", "bisquare"),
  simple_outlier_fraction = NULL,
  simple_outlier_thres = NULL,
  robust_tuning_method = c("M", "S"),
  previous_imp_weights = NULL
)

Arguments

values

a vector that may contains outliers. May be weights (inverse of pseudo-variance) or residuals(?!)

robust_method

If method = "none", then all subjects are included in the coefficient estimation. If method = "simple", then the outliers defined by simple_outlier_fraction or simple_outlier_thres are exlucded from the current iteration of coefficient estimation. If method = "huber", huber weights are assigned to each subject based on the “pseudo-residuals” rr.

simple_outlier_fraction

a value between 0 and 1, indicating the fraction of outliers: we define the simple_outlier_fraction percent of subjects with the highest absolute values (inverse of pseudo-variance) as outliers. By default, simple_outlier_fraction=NULL does not set any outlier fraction. Either simple_outlier_fraction or simple_outlier_thres should be specified when robust_method = "simple".

simple_outlier_thres

a real value, indicating the outliers whose inverse of pseudo-variance exceed this threshold to be removed from the current iteration. Either simple_outlier_fraction or simple_outlier_thres should be specified when robust_method = "simple".

robust_tuning_method

If robust_tuning_method = "M", M-estimation is performed. If robust_tuning_method = "S", S-estimation is performed. This argument specifies the tuning method, i.e., the method for defining outliers in each iteration, when applying Huber or Bisquare reweighting method (robust_method = "huber" or robust_method = "bisquare").

previous_imp_weights

importance weights in the previous iteration, used to calculate the importance weights in the current iteration for the S-estimation. If previous_imp_weights is NULL, it indicates the first iteration, and thus S-estimation need to be initialized.

Value

imp_weights An vector of length n, the importance weights. If a subject is to be removed from the current iteration, its importance weights is 0.