Update each single effect at a time

Update each single effect at a time

Usage

update_each_effect(
  X,
  y,
  gs,
  model,
  estimate_prior_variance = FALSE,
  estimate_prior_method = "optim",
  check_null_threshold = 0,
  abnormal_proportion = 0.5,
  robust_estimation = FALSE,
  robust_method = "simple",
  simple_outlier_fraction = 0.01,
  simple_outlier_thres = NULL,
  robust_tuning_method = NULL
)

Arguments

X

an n by p matrix of regression variables.

y

an n vector of response variable.

gs

a GSuSiE fit.

model

a list containing functions required for the specific generalized linear model.

estimate_prior_variance

Boolean. If TRUE, the coefficient prior variance is estimated and coef_prior_variance is then used as an initial value for the optimization (if provided). If FALSE, the prior variance for each of the maxL effects is fixed and determined by the value supplied to coef_prior_variance.

estimate_prior_method

The method used for estimating prior variance. When estimate_prior_method="simple" is used, the likelihood at the specified prior variance is compared to the likelihood at a variance of zero, and the setting with the larger likelihood is retained.

check_null_threshold

When the prior variance is estimated, compare the estimated with the null, and set the prior variance to zero unless the log-likelihood using the estimate is larger by this threshold amount.

abnormal_proportion

a value between 0 and 1. If the number of detected abnormal subjects exceeds \(abnormal_proportion * nrow(X)\), stop fitting the model.

robust_estimation

If robust_estimation=TRUE, robust estimation method is applied on the coefficient estimation.

robust_method

If robust_method= "simple", then the outliers defined by simple_outlier_fraction or simple_outlier_thres are simply removed in each iteration. If robust_method="huber", Huber weights are additionally multiplied to the pseudo-weights in each iteration. If robust_method = "bisquare", (Tukey's) bisquare weights are additionally multipled to the pseudo-weights in each iteration.

simple_outlier_fraction

a value between 0 and 1, indicating the fraction of outliers: we define the simple_outlier_fraction percent of subjects with the highest absolute values (inverse of pseudo-variance) as outliers. By default, simple_outlier_fraction=NULL does not set any outlier fraction. Either simple_outlier_fraction or simple_outlier_thres should be specified when robust_method = "simple".

simple_outlier_thres

a real value, indicating the outliers whose inverse of pseudo-variance exceed this threshold to be removed from the current iteration. Either simple_outlier_fraction or simple_outlier_thres should be specified when robust_method = "simple".

robust_tuning_method

If robust_tuning_method = "M", M-estimation is performed. If robust_tuning_method = "S", S-estimation is performed. This argument specifies the tuning method, i.e., the method for defining outliers in each iteration, when applying Huber or Bisquare reweighting method (robust_method = "huber" or robust_method = "bisquare").